Refactor api into core, breaking dependency of fedimint-server on mint-client
