+++ client/cli/src/main.rs
@@ -18,12 +18,13 @@ use fedimint_api::module::registry::ModuleDecoderRegistry;
 use fedimint_api::module::DynModuleGen;
 use fedimint_api::task::TaskGroup;
 use fedimint_api::{Amount, OutPoint, TieredMulti, TransactionId};
-use fedimint_core::config::load_from_file;
-use fedimint_mint::common::MintDecoder;
-use fedimint_mint::MintGen;
-use mint_client::api::{
+use fedimint_core::api::{
     FederationApiExt, GlobalFederationApi, IFederationApi, WsFederationApi, WsFederationConnect,
 };
+use fedimint_core::config::load_from_file;
+use fedimint_core::query::EventuallyConsistent;
+use fedimint_mint::common::MintDecoder;
+use fedimint_mint::MintGen;
 use mint_client::mint::SpendableNote;
 use mint_client::modules::ln::common::LightningDecoder;
 use mint_client::modules::ln::contracts::ContractId;
@@ -31,7 +32,6 @@ use mint_client::modules::ln::LightningGen;
 use mint_client::modules::wallet::common::WalletDecoder;
 use mint_client::modules::wallet::txoproof::TxOutProof;
 use mint_client::modules::wallet::WalletGen;
-use mint_client::query::EventuallyConsistent;
 use mint_client::utils::{
     from_hex, parse_bitcoin_amount, parse_ecash, parse_fedimint_amount, parse_node_pub_key,
     serialize_ecash,
