+++ client/client-lib/src/api/fake.rs
@@ -6,15 +6,14 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use fedimint_api::PeerId;
+use fedimint_core::api::IFederationApi;
+use fedimint_core::api::JsonRpcResult;
 use futures::Future;
 use serde;
 use serde::Serialize;
 use serde_json::Value;
 use tracing::info;
 use tracing::warn;
-
-use crate::api::IFederationApi;
-use crate::api::JsonRpcResult;
 
 #[allow(clippy::type_complexity)]
 type Handler<State> = Pin<
