+++ fedimint-core/Cargo.toml
@@ -21,14 +21,27 @@ bitcoin = "0.29.2"
 itertools = "0.10.5"
 fedimint-api = { path = "../fedimint-api" }
 rand = "0.8"
+jsonrpsee-types = "0.16.0"
+jsonrpsee-core = { version = "0.16.2", features = [ "client" ] }
 secp256k1-zkp = { version = "0.7.0", features = [ "global-context", "bitcoin_hashes" ] }
 serde = { version = "1.0.149", features = [ "derive" ] }
 serde_json = "1.0.91"
 sha3 = "0.10.5"
 tbs = { path = "../crypto/tbs" }
+tokio = { version = "1.24.2", features = ["sync"] }
 thiserror = "1.0.37"
 tracing ="0.1.37"
 threshold_crypto = { git = "https://github.com/jkitman/threshold_crypto", branch = "upgrade-threshold-crypto-libs" }
 bitcoin_hashes = "0.11.0"
 url = { version = "2.3.1", features = ["serde"] }
 
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+jsonrpsee-ws-client = "0.16.2"
+
+[target.'cfg(target_family = "wasm")'.dependencies]
+jsonrpsee-wasm-client = "0.16.0"
+
+[dev-dependencies]
+test-log = { version = "0.2", features = [ "trace" ], default-features = false }
+once_cell = "1.16.0"
+tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
