+++ fedimint-server/Cargo.toml
@@ -24,7 +24,6 @@ futures = "0.3.24"
 impl-tools = "0.6.1"
 itertools = "0.10.5"
 jsonrpsee = { version = "0.16.2", features = ["server"] }
-mint-client = { path = "../client/client-lib" }
 fedimint-api = { path = "../fedimint-api" }
 fedimint-core = { path = "../fedimint-core" }
 rand = "0.8"
