+++ fedimint-server/src/lib.rs
@@ -14,6 +14,8 @@ use fedimint_api::module::registry::ModuleDecoderRegistry;
 use fedimint_api::net::peers::PeerConnections;
 use fedimint_api::task::{sleep, TaskGroup, TaskHandle};
 use fedimint_api::{NumPeers, PeerId};
+use fedimint_core::api::WsFederationApi;
+use fedimint_core::api::{DynFederationApi, GlobalFederationApi};
 use fedimint_core::epoch::{
     ConsensusItem, EpochVerifyError, SerdeConsensusItem, SignedEpochOutcome,
 };
@@ -25,8 +27,6 @@ use futures::StreamExt;
 use hbbft::honey_badger::{Batch, HoneyBadger, Message, Step};
 use hbbft::{Epoched, NetworkInfo, Target};
 use itertools::Itertools;
-use mint_client::api::WsFederationApi;
-use mint_client::api::{DynFederationApi, GlobalFederationApi};
 use rand::rngs::OsRng;
 use rand::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
