+++ fedimintd/src/lib.rs
@@ -5,10 +5,10 @@ use anyhow::format_err;
 use bitcoin::Network;
 use fedimint_api::config::{ConfigGenParams, ModuleGenRegistry};
 use fedimint_api::{Amount, Tiered};
+use fedimint_core::api::WsFederationConnect;
 use fedimint_mint::MintGenParams;
 use fedimint_server::config::ServerConfig;
 use fedimint_wallet::WalletGenParams;
-use mint_client::api::WsFederationConnect;
 use ring::aead::LessSafeKey;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
