+++ fedimintd/src/ui.rs
@@ -16,9 +16,9 @@ use fedimint_api::bitcoin_rpc::BitcoindRpcBackend;
 use fedimint_api::config::{ClientConfig, ModuleGenRegistry};
 use fedimint_api::task::TaskGroup;
 use fedimint_api::Amount;
+use fedimint_core::api::WsFederationConnect;
 use fedimint_core::util::SanitizedUrl;
 use http::StatusCode;
-use mint_client::api::WsFederationConnect;
 use qrcode_generator::QrCodeEcc;
 use serde::Deserialize;
 use tokio::select;
