+++ gateway/ln-gateway/src/gatewayd/gateway.rs
@@ -14,8 +14,9 @@ use fedimint_api::{
     task::TaskGroup,
     Amount, TransactionId,
 };
+use fedimint_server::api::WsFederationConnect;
 use mint_client::modules::ln::{contracts::Preimage, route_hints::RouteHint};
-use mint_client::{api::WsFederationConnect, ln::PayInvoicePayload, GatewayClient};
+use mint_client::{ln::PayInvoicePayload, GatewayClient};
 use tokio::sync::{mpsc, Mutex};
 use tracing::{error, info, warn};
